Refactor polls into multiple files
