+++ src/polls/pollQuestionUi.ts
@@ -1,0 +1,27 @@
+import * as ui from 'dcl-ui-toolkit'
+
+export function createPollQuestionUi(pollQuestion: string, options: string[] = ['Yeah', 'Nope']): ui.CustomPrompt {
+    const prompt = ui.createComponent(ui.CustomPrompt, { style: ui.PromptStyles.DARKSLANTED })
+    prompt.addText({
+      value: pollQuestion,
+      size: 30,
+      yPosition: 170,
+      xPosition: -100
+    })
+  
+    const buttonSpacing = 50
+    const startY = ((options.length - 1) / 2) * buttonSpacing
+  
+    options.forEach((option, index) => {
+      prompt.addButton({
+        text: option,
+        yPosition: startY - index * buttonSpacing,
+        xPosition: -100,
+        onMouseDown: () => { prompt.hide() }
+      })
+    })
+  
+    prompt.show()
+
+    return prompt
+  }
